requirements.txt
google-api-python-client
python-dotenv
youtube-transcript-api
openai
markdown
ebooklib
requests
